Make the mypy fixture available to the whole test suite
Also make mypy tests automatically slow.

See #130 for other cases where a mypy fixture would be useful.
